Fix compile error in windows

+++ src/iarray_container.c
@@ -1056,7 +1056,7 @@ INA_API(ina_rc_t) iarray_container_equal(iarray_container_t *a, iarray_container
                 for (int64_t i = 0; i < val_a.block_size; ++i) {
                     int64_t diff = ((int64_t *)val_a.block_pointer)[i] - ((int64_t *)val_b.block_pointer)[i];
                     if (diff != 0) {
-                        printf("%lld, %lld (diff: %lld)\n", ((long long *)val_a.block_pointer)[i], ((long long *)val_b.block_pointer)[i], diff);
+                        printf("%lld, %lld (diff: %lld)\n", ((long long *)val_a.block_pointer)[i], ((long long *)val_b.block_pointer)[i], (long long)diff);
                         IARRAY_TRACE1(iarray.error, "Values are different");
                         return INA_ERROR(IARRAY_ERR_ASSERTION_FAILED);
                     }
